+++ silly_kicks/vaep/labels.py
@@ -142,3 +142,43 @@ def goal_from_shot(actions: pd.DataFrame) -> pd.DataFrame:
     goals = actions["type_name"].str.contains("shot") & (actions["result_id"] == spadl.result_id["success"])
 
     return pd.DataFrame(goals, columns=["goal_from_shot"])
+
+
+def save_from_shot(actions: pd.DataFrame) -> pd.DataFrame:
+    """Determine whether the goalkeeper saved the current shot.
+
+    This label can be used to train an Expected Saves (xS) model.
+
+    Parameters
+    ----------
+    actions : pd.DataFrame
+        The actions of a game.
+
+    Returns
+    -------
+    pd.DataFrame
+        A dataframe with a column 'save_from_shot' and a row for each action
+        set to True if the action is a keeper save; otherwise False.
+    """
+    saves = actions["type_name"].str.contains("keeper_save") & (actions["result_id"] == spadl.result_id["success"])
+    return pd.DataFrame(saves, columns=["save_from_shot"])
+
+
+def claim_from_cross(actions: pd.DataFrame) -> pd.DataFrame:
+    """Determine whether the goalkeeper claimed the current cross.
+
+    This label can be used to train an Expected Claims (xC) model.
+
+    Parameters
+    ----------
+    actions : pd.DataFrame
+        The actions of a game.
+
+    Returns
+    -------
+    pd.DataFrame
+        A dataframe with a column 'claim_from_cross' and a row for each action
+        set to True if the action is a keeper claim; otherwise False.
+    """
+    claims = actions["type_name"].str.contains("keeper_claim") & (actions["result_id"] == spadl.result_id["success"])
+    return pd.DataFrame(claims, columns=["claim_from_cross"])